ai-coder/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2
python-multipart==0.0.6
redis==5.0.1
python-dotenv==1.0.0
# RAG Dependencies
qdrant-client==1.7.0
neo4j==5.15.0
openai==1.6.1
tiktoken==0.5.2
# ============================================================================
# NEW EMBEDDINGS DEPENDENCIES
# ============================================================================

# HTTP Client (already have requests, just ensure aiohttp is added)
aiohttp==3.9.1

# Local Embeddings (Fallback)
sentence-transformers==2.2.2
torch==2.2.0
transformers==4.35.0

# Google Gemini (Fallback)
google-generativeai==0.3.2

# Caching & Compression
msgpack==1.0.7
orjson==3.9.10

